backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cors==1.0.1
filelock==3.12.2
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gTTS==2.3.2
hume==0.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.0
openai==0.27.8
opencv-python==4.7.0.72
playsound==1.3.0
pvrecorder==1.1.2
pynput==1.7.6
pyobjc-core==9.2
pyobjc-framework-ApplicationServices==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
PySocks==1.7.1
python-dotenv==1.0.0
pywatchman==1.4.1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
tldextract==3.4.4
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
websockets==10.4
Werkzeug==2.3.6
whispercpp @ git+https://github.com/aarnphm/whispercpp.git@69cadf0d4a2fc8b640df80313e53dc7a208d469a
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
